RF: move progressbars imports into a separate module to delay their imports until used
See https://github.com/tqdm/tqdm/issues/176  for one of the reasons but this one
is also an attempt to mitigate a weird tests failure due to import of tqdm, see eg

http://smaug.datalad.org:8020/builders/datalad-pr-docker-dl-nd90/builds/785/steps/nosetests/logs/stdio
